threads WithServerSideProps through mapFields in create first user view
