+++ packages/ui/src/providers/ComponentMap/index.tsx
@@ -17,6 +17,8 @@ export type IComponentMapContext = {
   }) => MappedField | undefined
 }
 
+export { WithServerSideProps } from './buildComponentMap/WithServerSideProps.js'
+
 const ComponentMapContext = createContext<IComponentMapContext>({} as IComponentMapContext)
 
 export const ComponentMapProvider: React.FC<{
